chore: table config
