+++ src/apps/Handover/Garden/utility/gardenSetup.ts
@@ -20,67 +20,6 @@ export const fieldSettings: FieldSettings<HandoverPackage, ExtendedGardenFields>
     priority2: { label: 'Commissioning Priority 2' },
     priority3: { label: 'Commissioning Priority 3' },
 };
-export const hiddenColumns: (keyof HandoverPackage)[] = [
-    'siteCode',
-    'projectIdentifier',
-    'projectDescription',
-    'progress',
-    'area',
-    'priority1Description',
-    'priority2Description',
-    'priority3Description',
-    'actualStartDate',
-    'plannedFinishDate',
-    'actualFinishDate',
-    'url',
-    'id',
-    'hasUnsignedActions',
-    'forecastFinishDate',
-    'volume',
-    'forecastTacDate',
-    'mcPkgsCount',
-    'mcPkgsRFCCShippedCount',
-    'mcPkgsRFCCSigned',
-    'mcPkgsRFOCShipped',
-    'mcPkgsRFOCSigned',
-    'tacIsAccepted',
-    'tacIsShipped',
-    'tacIsRejected',
-    'plannedTacDate',
-    'mcDisciplineCodes',
-    'mcDisciplines',
-    'isSubsea',
-    'isDemolition',
-    'demolitionPlannedStartDate',
-    'demolitionForecastStartDate',
-    'demolitionActualStartDate',
-    'demolitionPlannedFinishDate',
-    'demolitionForecastFinishDate',
-    'demolitionActualFinishDate',
-    'createdDate',
-    'remark',
-    'rfocIsShipped',
-    'rfocIsAccepted',
-    'rfocIsRejected',
-    'rfccIsShipped',
-    'rfccIsAccepted',
-    'rfccIsRejected',
-    'subSystem',
-    'isReadyForStartup',
-    'isInOperation',
-    'hasMaintenanceProgram',
-    'hasYellowLineMarkup',
-    'hasBlueLineMarkup',
-    'yellowLineStatus',
-    'blueLineStatus',
-    'hasOperationAgreement',
-    'demolitionDCCAcceptedDate',
-    'demolitionRFRCShippedDate',
-    'tacActualDate',
-    'rfccShippedDate',
-    'rfocShippedDate',
-    'rowKey',
-];
 
 export const removedFilterOptions: (keyof HandoverPackage)[] = [
     'createdDate',
